requirements.txt
amqp==2.5.2
appdirs==1.4.3
asgiref==3.2.3
attrs==19.3.0
backcall==0.1.0
billiard==3.6.2.0
black==19.10b0
celery==4.4.0
Click==7.0
colorama==0.4.3
decorator==4.4.1
Django==3.0.3
django-cors-headers==3.2.1
django-debug-toolbar==2.2
importlib-metadata==1.5.0
ipython==7.12.0
ipython-genutils==0.2.0
jedi==0.16.0
kombu==4.6.7
parso==0.6.1
pathspec==0.7.0
pexpect==4.8.0
pickleshare==0.7.5
prompt-toolkit==3.0.3
ptyprocess==0.6.0
Pygments==2.5.2
pytz==2019.3
regex==2020.1.8
six==1.14.0
sqlparse==0.3.0
toml==0.10.0
traitlets==4.3.3
typed-ast==1.4.1
vine==1.3.0
wcwidth==0.1.8
zipp==2.2.0
uwsgi
